# [release-3.5] fix the potential data loss for clusters with only one member