@serathius 2022-09-06
Motivation for benchmark was to verify that performance regression is within expected range (~5% based on your tests). Even if I provided environment I don't expect that could be used to reproduce the results as I don't expect most people to have same CPU/Mem setup. 

However we are right about different environments providing different results, for public announcements we should try to diversify our testing environments to have more accurate results. Instead of fixed number 4.5% we could provide a range (for example) 3-6% which could be reflect best and worst case scenarios we seen in different environments.